requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
jinja2==2.11.3
    # via -r requirements.in
markupsafe==2.1.1
    # via jinja2
pillow==9.2.0
    # via -r requirements.in
tqdm==4.54.0
    # via -r requirements.in
unidecode==1.2.0
    # via -r requirements.in
